调整：移除未同步修改的全局提示
